refactor(folder structure) : move server's class
move the class with the server role and server's util class to the server folder
