+++ src/server/Server.cpp
@@ -11,7 +11,7 @@
 /* ************************************************************************** */
 
 #include "Server.hpp"
-#include "Library.hpp"
+#include "../Library.hpp"
 
 Server::Server() : serverFd(0) {}
