+++ src/server/ServerConfig.hpp
@@ -13,7 +13,7 @@
 #ifndef SERVER_CONFIG_HPP
 # define SERVER_CONFIG_HPP
 
-#include "Library.hpp"
+#include "../Library.hpp"
 #include "Location.hpp"
 
 class ServerConfig {
